% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pguDMwR.R
\name{centralValue}
\alias{centralValue}
\title{centralValue}
\usage{
centralValue(x, ws = NULL)
}
\arguments{
\item{x}{variable}

\item{ws}{weights}
}
\value{
central value
}
\description{
Returns the central value of a variable.
}
\details{
Function that obtains a statistic of centrality of a
variable, given a sample of values.
If the variable is numeric it returns de median, if it
is a factor it returns the mode. In other cases it
tries to convert to a factor and then returns the mode.
Taken from:
https://github.com/ltorgo/DMwR2/
}
\examples{
centralValue(x = seq(1,10,1))

}
\author{
Luis Torgo
}
