% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dLogLikelihood.R
\name{dLogLikelihood}
\alias{dLogLikelihood}
\title{dLogLikelihood}
\usage{
dLogLikelihood(x = "numeric", mu = "numeric", sigma = "numeric", log = TRUE)
}
\arguments{
\item{x}{The x-value.
(numeric)}

\item{mu}{The expextation value.
(numeric)}

\item{sigma}{The standard deviation.
(numeric)}

\item{log}{Indicate to calculate the log likelihood
(logical)}
}
\value{
The logLikelihood.
(numeric)
}
\description{
Calculates the logLikelihood.
}
\examples{
y <- dLogLikelihood(x=5, mu=0.0, sigma=1.0)

}
\author{
Sebastian Malkusch, \email{malkusch@med.uni-frankfurt.de}
}
