% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importDataSet.R
\name{importDataSet}
\alias{importDataSet}
\title{Imports a dataset to the shiny pguIMP web interface.}
\usage{
importDataSet(obj = "pgu.file")
}
\arguments{
\item{obj}{Instance of type R6 and pguIMP::pgu.file}
}
\value{
A data frame
(tibble::tibble)
}
\description{
Extracts import options from a pguIMP::file instance
and imports the desired record based on the passed information.
}
