% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pguDMwR.R
\name{knnImputation}
\alias{knnImputation}
\title{knnImputation}
\usage{
knnImputation(data, k = 10, scale = TRUE, meth = "weighAvg", distData = NULL)
}
\arguments{
\item{data}{data frame containing missing values}

\item{k}{number of nearest neighbors}

\item{scale}{Indicates if data should be scaled}

\item{meth}{Method for estimating the missing value}

\item{distData}{Distance to the case}
}
\value{
cleaned data
}
\description{
Imputes missings using kNN.
}
\details{
Function that fills in all unknowns using the k Nearest
Neighbours of each case with unknows.
By default it uses the values of the neighbours and
obtains an weighted (by the distance to the case) average
of their values to fill in the unknows.
If meth='median' it uses the median/most frequent value,
instead.
Taken from
https://github.com/ltorgo/DMwR2/
}
\examples{
centralValue(x = seq(1,10,1))

}
\author{
Luis Torgo
}
