% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalDistribution.R
\name{normalDistribution}
\alias{normalDistribution}
\title{normalDistribution}
\usage{
normalDistribution(x = "numeric", mu = "numeric", sigma = "numeric")
}
\arguments{
\item{x}{The x-value.
(numeric)}

\item{mu}{The expextation value.
(numeric)}

\item{sigma}{The standard deviation.
(numeric)}
}
\value{
The probability of event x.
(numeric)
}
\description{
Represents a normal distribution.
}
\details{
Calculates p(x | mu, sigma). Where p is a normal distribution
with user defined expectation value \code{mu}
and standard deviation \code{sigma}.
}
\examples{
y <- normalDistribution(x=5, mu=0.0, sigma=1.0)

}
\author{
Sebastian Malkusch, \email{malkusch@med.uni-frankfurt.de}
}
