% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pguCorrValidator.R
\name{pgu.corrValidator}
\alias{pgu.corrValidator}
\title{pgu.corrValidator}
\format{
\link[R6:R6Class]{R6::R6Class} object.
}
\description{
A class that performs pairwise correlation on the pguIMP data set.
}
\details{

}
\section{Construction}{

x <- pguIMP::pgu.correlator$new(data)
}

\examples{

## ------------------------------------------------
## Method `pgu.corrValidator$new`
## ------------------------------------------------

x <- tibble::tibble()
y <- impute(x)
obj <- pguIMP::pgu.corrValidator$new(org_df = x, imp_df = y)

## ------------------------------------------------
## Method `pgu.corrValidator$print`
## ------------------------------------------------

pgu.corrValidator$print()
print(pgu.corrValidator)

## ------------------------------------------------
## Method `pgu.corrValidator$reset`
## ------------------------------------------------

pgu.corrValidator$reset()

## ------------------------------------------------
## Method `pgu.corrValidator$resetMatrix`
## ------------------------------------------------

matrix <- pgu.corrValidator$resetMatrix(value)

## ------------------------------------------------
## Method `pgu.corrValidator$flattenMatrix`
## ------------------------------------------------

pgu.corrValidatior$flattenMatrix()

## ------------------------------------------------
## Method `pgu.corrValidator$validate`
## ------------------------------------------------

pgu.corrValidator$validate(org_df, imp_df)

## ------------------------------------------------
## Method `pgu.corrValidator$summary`
## ------------------------------------------------

df <- pgu.corrValidator$summary()

## ------------------------------------------------
## Method `pgu.corrValidator$correlationScatterPlot`
## ------------------------------------------------

pgu.corrValidator$correlationScatterPlot()

## ------------------------------------------------
## Method `pgu.corrValidator$correlationBoxPlot`
## ------------------------------------------------

pgu.corrValidator$correlationBoxPlot()
}
\author{
Sebastian Malkusch, \email{malkusch@med.uni-frankfurt.de}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{featureNames}}{Returns the instance variable featureNames.
(character)}

\item{\code{orgR_mat}}{Returns the instance variable orgR_mat.
(matrix)}

\item{\code{impR_mat}}{Returns the instance variable impR_mat.
(matrix)}

\item{\code{orgP_mat}}{Returns the instance variable orgP_mat.
(matrix)}

\item{\code{impP_mat}}{Returns the instance variable impP_mat.
(matrix)}

\item{\code{corr_df}}{Returns the instance variable corr_df.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{pgu.corrValidator$new()}}
\item \href{#method-finalize}{\code{pgu.corrValidator$finalize()}}
\item \href{#method-print}{\code{pgu.corrValidator$print()}}
\item \href{#method-reset}{\code{pgu.corrValidator$reset()}}
\item \href{#method-resetMatrix}{\code{pgu.corrValidator$resetMatrix()}}
\item \href{#method-flattenMatrix}{\code{pgu.corrValidator$flattenMatrix()}}
\item \href{#method-validate}{\code{pgu.corrValidator$validate()}}
\item \href{#method-summary}{\code{pgu.corrValidator$summary()}}
\item \href{#method-correlationScatterPlot}{\code{pgu.corrValidator$correlationScatterPlot()}}
\item \href{#method-correlationBoxPlot}{\code{pgu.corrValidator$correlationBoxPlot()}}
\item \href{#method-clone}{\code{pgu.corrValidator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates and returns a new \code{pgu.corrValidator} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.corrValidator$new(org_df = "tbl_df", imp_df = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_df}}{The original data to be abalyzed.
(tibble::tibble)}

\item{\code{imp_df}}{The imputed version of the org_df data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{pgu.corrValidator} object.
(pguIMP::pgu.corrValidator)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x <- tibble::tibble()
y <- impute(x)
obj <- pguIMP::pgu.corrValidator$new(org_df = x, imp_df = y)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Clears the heap and
indicates if instance of \code{pgu.corrValidator} is removed from heap.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.corrValidator$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints instance variables of a \code{pgu.corrValidator} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.corrValidator$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
string
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{pgu.corrValidator$print()
print(pgu.corrValidator)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\if{latex}{\out{\hypertarget{method-reset}{}}}
\subsection{Method \code{reset()}}{
Resets the object \code{pgu.corrValidator} based on the instance variable featureNames..
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.corrValidator$reset()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{pgu.corrValidator$reset()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-resetMatrix"></a>}}
\if{latex}{\out{\hypertarget{method-resetMatrix}{}}}
\subsection{Method \code{resetMatrix()}}{
Creates a square matrix which dimension corresponds to the length
of the instance variable featureNames. The matrix entries are set to a distinct \code{value}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.corrValidator$resetMatrix(value = "numeric")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The value the matrix entries are set to.
(numeric)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A square matrix.
(matrix)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{matrix <- pgu.corrValidator$resetMatrix(value)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-flattenMatrix"></a>}}
\if{latex}{\out{\hypertarget{method-flattenMatrix}{}}}
\subsection{Method \code{flattenMatrix()}}{
Flattens the results transforms them into a dataframe
and stores it into the instance variable corr_df.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.corrValidator$flattenMatrix()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{pgu.corrValidatior$flattenMatrix()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
Runs the corraltion analysis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.corrValidator$validate(org_df = "tbl_df", imp_df = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org_df}}{Adataframe comprising the original data.
(tibble::tibble)}

\item{\code{imp_df}}{Adataframe comprising the imputed data.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{pgu.corrValidator$validate(org_df, imp_df)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\if{latex}{\out{\hypertarget{method-summary}{}}}
\subsection{Method \code{summary()}}{
Summary of the correlation deviation distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.corrValidator$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A dataframe comprising the summary of the correlation deviation distribution.
(tibble::tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{df <- pgu.corrValidator$summary()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-correlationScatterPlot"></a>}}
\if{latex}{\out{\hypertarget{method-correlationScatterPlot}{}}}
\subsection{Method \code{correlationScatterPlot()}}{
Plots the correlation analysis results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.corrValidator$correlationScatterPlot()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{pgu.corrValidator$correlationScatterPlot()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-correlationBoxPlot"></a>}}
\if{latex}{\out{\hypertarget{method-correlationBoxPlot}{}}}
\subsection{Method \code{correlationBoxPlot()}}{
Plots the correlation analysis results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.corrValidator$correlationBoxPlot()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{pgu.corrValidator$correlationBoxPlot()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.corrValidator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
