% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pguData.R
\name{pgu.data}
\alias{pgu.data}
\title{pgu.data}
\format{
\link[R6:R6Class]{R6::R6Class} object.
}
\description{
Handles the pguIMP dataset.
}
\details{
Stores the pguIMP dataset as instance variable and keeps track of the attributes of interest.
Provides additionally fast access to several statistical information about the data set.
}
\section{Construction}{

x <- pguIMP::pgu.data$new(data)
}

\examples{

## ------------------------------------------------
## Method `pgu.data$new`
## ------------------------------------------------

y <- tibble::tibble()
x <- pguIMP:pgu.data$new(data = y)

## ------------------------------------------------
## Method `pgu.data$print`
## ------------------------------------------------

x$print()
print(x)

## ------------------------------------------------
## Method `pgu.data$resetActiveFeatureNames`
## ------------------------------------------------

x$resetActiveFeatureNames()

## ------------------------------------------------
## Method `pgu.data$featureIdx`
## ------------------------------------------------

idx <- x$featureIdx(feature = "infected")

## ------------------------------------------------
## Method `pgu.data$numericData`
## ------------------------------------------------

y <- x$numericData()

## ------------------------------------------------
## Method `pgu.data$nonNumericData`
## ------------------------------------------------

y <- x$nonNumericData()

## ------------------------------------------------
## Method `pgu.data$dataInformation`
## ------------------------------------------------

x$dataInformation() \%>\%
 print()

## ------------------------------------------------
## Method `pgu.data$summarizeNumeric`
## ------------------------------------------------

x$summarizeNumeric(val = c(1,2,3,4,5))

## ------------------------------------------------
## Method `pgu.data$dataStatistics`
## ------------------------------------------------

x$dataStatistics() \%>\%
 print()

## ------------------------------------------------
## Method `pgu.data$reducedDataStatistics`
## ------------------------------------------------

x$reducedDataStatistics() \%>\%
 print()

## ------------------------------------------------
## Method `pgu.data$missings`
## ------------------------------------------------

x$misings() \%>\%
 print()
}
\author{
Sebastian Malkusch, \email{malkusch@med.uni-frankfurt.de}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{rawData}}{Returns the instance variable rawData
(tibble::tibble)}

\item{\code{setRawData}}{Sets the instance variable rawData
(tibble::tibble)}

\item{\code{featureNames}}{Returns the instance variable featureNames
(character)}

\item{\code{numericFeatureNames}}{Returns the insatnce variable numericFeatureNames
(character)}

\item{\code{nonNumericFeatureNames}}{Returns the instance variable nonNumericFeatureNames
(character)}

\item{\code{activeFeatureNames}}{Returns the instance variable activeFeatureNames
(character)}

\item{\code{setActiveFeatureNames}}{Sets the instance variable activeFeatureNames
(character)}

\item{\code{abscissa}}{Returns the instance variable abscissa
(character)}

\item{\code{setAbscissa}}{Sets the instance variable abscissa
(character)}

\item{\code{ordinate}}{Returns the instance variable ordinate
(character)}

\item{\code{setOrdinate}}{Sets the instance variable ordinate
(character)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{pgu.data$new()}}
\item \href{#method-finalize}{\code{pgu.data$finalize()}}
\item \href{#method-print}{\code{pgu.data$print()}}
\item \href{#method-resetActiveFeatureNames}{\code{pgu.data$resetActiveFeatureNames()}}
\item \href{#method-featureIdx}{\code{pgu.data$featureIdx()}}
\item \href{#method-numericData}{\code{pgu.data$numericData()}}
\item \href{#method-nonNumericData}{\code{pgu.data$nonNumericData()}}
\item \href{#method-dataInformation}{\code{pgu.data$dataInformation()}}
\item \href{#method-summarizeNumeric}{\code{pgu.data$summarizeNumeric()}}
\item \href{#method-dataStatistics}{\code{pgu.data$dataStatistics()}}
\item \href{#method-reducedDataStatistics}{\code{pgu.data$reducedDataStatistics()}}
\item \href{#method-missings}{\code{pgu.data$missings()}}
\item \href{#method-clone}{\code{pgu.data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates and returns a new \code{pgu.data} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.data$new(data = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data to be analyzed.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{pgu.data} object.
(pguIMP::pgu.data)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{y <- tibble::tibble()
x <- pguIMP:pgu.data$new(data = y)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Clears the heap and
indicates that instance of \code{pgu.data} is removed from heap.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.data$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints instance variables of a \code{pgu.data} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.data$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
string
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$print()
print(x)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-resetActiveFeatureNames"></a>}}
\if{latex}{\out{\hypertarget{method-resetActiveFeatureNames}{}}}
\subsection{Method \code{resetActiveFeatureNames()}}{
Sets all attribute feature names as active.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.data$resetActiveFeatureNames()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$resetActiveFeatureNames()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featureIdx"></a>}}
\if{latex}{\out{\hypertarget{method-featureIdx}{}}}
\subsection{Method \code{featureIdx()}}{
Returns the index of an attributes feature name within the object's instance variable \code{featureNames}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.data$featureIdx(feature = "character")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature}}{Attribute's name.
(character)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Index of attribute's feature name in dataframe
(numeric)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{idx <- x$featureIdx(feature = "infected")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-numericData"></a>}}
\if{latex}{\out{\hypertarget{method-numericData}{}}}
\subsection{Method \code{numericData()}}{
Returns a copy of the instance variable rawData comprising solely the numeric attributes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.data$numericData()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data frame
(tibble::tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{y <- x$numericData()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nonNumericData"></a>}}
\if{latex}{\out{\hypertarget{method-nonNumericData}{}}}
\subsection{Method \code{nonNumericData()}}{
Returns a copy of the instance variable rawData comprising solely the attributes of non numeric types.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.data$nonNumericData()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data frame
(tibble::tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{y <- x$nonNumericData()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataInformation"></a>}}
\if{latex}{\out{\hypertarget{method-dataInformation}{}}}
\subsection{Method \code{dataInformation()}}{
Summarizes information on the instance variable rawData
and retruns it in form of a compact data frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.data$dataInformation()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Information data frame
(tibble::tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$dataInformation() \%>\%
 print()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarizeNumeric"></a>}}
\if{latex}{\out{\hypertarget{method-summarizeNumeric}{}}}
\subsection{Method \code{summarizeNumeric()}}{
Summarizes a vector of numerics and returns summary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.data$summarizeNumeric(val = "numeric")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Vector of numerics to be summarized.
(numeric)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Summary
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$summarizeNumeric(val = c(1,2,3,4,5))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataStatistics"></a>}}
\if{latex}{\out{\hypertarget{method-dataStatistics}{}}}
\subsection{Method \code{dataStatistics()}}{
Iterativly calls the class' function summarizeNumeric on all numeric attributes
of the instance variable rawData and returns the result in form of a data frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.data$dataStatistics()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Summary
(tibble:tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$dataStatistics() \%>\%
 print()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reducedDataStatistics"></a>}}
\if{latex}{\out{\hypertarget{method-reducedDataStatistics}{}}}
\subsection{Method \code{reducedDataStatistics()}}{
Calls the class' function dataStatistics
filters the result for the attributes feature names and arithmetric mean values.
and returns the result in form of a data frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.data$reducedDataStatistics()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Summary
(tibble:tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$reducedDataStatistics() \%>\%
 print()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-missings"></a>}}
\if{latex}{\out{\hypertarget{method-missings}{}}}
\subsection{Method \code{missings()}}{
Calls the class' function dataStatistics
filters the result for the attributes feature names and information about missing values.
and returns the result in form of a data frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.data$missings()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Summary
(tibble:tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$misings() \%>\%
 print()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
