% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pguExplorer.R
\name{pgu.explorer}
\alias{pgu.explorer}
\title{pgu.explorer}
\format{
\link[R6:R6Class]{R6::R6Class} object.
}
\description{
Visual exploration of the pguIMP dataset.
}
\details{
Pariwise anlysis of attributes from the pguIMP dataset.
}
\section{Construction}{

x <- pguIMP::pgu.explorer$new(data)
}

\examples{

## ------------------------------------------------
## Method `pgu.explorer$new`
## ------------------------------------------------

y <- tibble:tibble()
x <- pguIMP:pgu.explorer$new(data = y)

## ------------------------------------------------
## Method `pgu.explorer$print`
## ------------------------------------------------

x$print()
print(x)

## ------------------------------------------------
## Method `pgu.explorer$reset`
## ------------------------------------------------

x$reset(abj, abs = "time", ord = "infected")

## ------------------------------------------------
## Method `pgu.explorer$abscissaIsNumeric`
## ------------------------------------------------

y <- x$abscissaIsNumeric()

## ------------------------------------------------
## Method `pgu.explorer$ordinateIsNumeric`
## ------------------------------------------------

y <- x$ordinateIsNumeric()

## ------------------------------------------------
## Method `pgu.explorer$scatterPlot`
## ------------------------------------------------

x$scatterPlot() \%>\%
 show()

## ------------------------------------------------
## Method `pgu.explorer$abscissaBarPlot`
## ------------------------------------------------

x$abscissaBarPlot() \%>\%
 show()

## ------------------------------------------------
## Method `pgu.explorer$abscissaBoxPlot`
## ------------------------------------------------

x$abscissaBoxPlot() \%>\%
 show()

## ------------------------------------------------
## Method `pgu.explorer$abscissaPlot`
## ------------------------------------------------

x$abscissaPlot() \%>\%
 show()

## ------------------------------------------------
## Method `pgu.explorer$ordinateBarPlot`
## ------------------------------------------------

x$ordinateBarPlot() \%>\%
 show()

## ------------------------------------------------
## Method `pgu.explorer$ordinateBoxPlot`
## ------------------------------------------------

x$ordinateBoxPlot() \%>\%
 show()

## ------------------------------------------------
## Method `pgu.explorer$ordinatePlot`
## ------------------------------------------------

x$ordinatePlot() \%>\%
 show()

## ------------------------------------------------
## Method `pgu.explorer$summarizeNumeric`
## ------------------------------------------------

x$summarizeNumeric(val) \%>\%
 print()

## ------------------------------------------------
## Method `pgu.explorer$abscissaStatistic`
## ------------------------------------------------

x$abscissaStatistic() \%>\%
 print()

## ------------------------------------------------
## Method `pgu.explorer$ordinateStatistic`
## ------------------------------------------------

x$ordinateStatistic() \%>\%
 print()
}
\author{
Sebastian Malkusch, \email{malkusch@med.uni-frankfurt.de}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{rawData}}{Returns the instance variable rawData
(tibble::tibble)}

\item{\code{setRawData}}{Sets the instance variable rawData
(tibble::tibble)}

\item{\code{abscissa}}{Returns the instance variable abscissa
(character)}

\item{\code{setAbscissa}}{Sets the instance variable abscissa
(character)}

\item{\code{ordinate}}{Returns the instance variable ordinate
(character)}

\item{\code{setOrdinate}}{Sets the instance variable ordinate
(character)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{pgu.explorer$new()}}
\item \href{#method-finalize}{\code{pgu.explorer$finalize()}}
\item \href{#method-print}{\code{pgu.explorer$print()}}
\item \href{#method-reset}{\code{pgu.explorer$reset()}}
\item \href{#method-abscissaIsNumeric}{\code{pgu.explorer$abscissaIsNumeric()}}
\item \href{#method-ordinateIsNumeric}{\code{pgu.explorer$ordinateIsNumeric()}}
\item \href{#method-scatterPlot}{\code{pgu.explorer$scatterPlot()}}
\item \href{#method-abscissaBarPlot}{\code{pgu.explorer$abscissaBarPlot()}}
\item \href{#method-abscissaBoxPlot}{\code{pgu.explorer$abscissaBoxPlot()}}
\item \href{#method-abscissaPlot}{\code{pgu.explorer$abscissaPlot()}}
\item \href{#method-ordinateBarPlot}{\code{pgu.explorer$ordinateBarPlot()}}
\item \href{#method-ordinateBoxPlot}{\code{pgu.explorer$ordinateBoxPlot()}}
\item \href{#method-ordinatePlot}{\code{pgu.explorer$ordinatePlot()}}
\item \href{#method-summarizeNumeric}{\code{pgu.explorer$summarizeNumeric()}}
\item \href{#method-abscissaStatistic}{\code{pgu.explorer$abscissaStatistic()}}
\item \href{#method-ordinateStatistic}{\code{pgu.explorer$ordinateStatistic()}}
\item \href{#method-clone}{\code{pgu.explorer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates and returns a new \code{pgu.explorer} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.explorer$new(obj = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{pgu.explorer} object.
(pguIMP::pgu.optimizer)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{y <- tibble:tibble()
x <- pguIMP:pgu.explorer$new(data = y)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Clears the heap and
indicates that instance of \code{pgu.explorer} is removed from heap.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.explorer$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints instance variables of a \code{pgu.explorer} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.explorer$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
string
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$print()
print(x)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\if{latex}{\out{\hypertarget{method-reset}{}}}
\subsection{Method \code{reset()}}{
Resets the instance of the pgu.explorer class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.explorer$reset(obj = "tbl_df", abs = "character", ord = "character")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
(tibble::tibble)}

\item{\code{abs}}{The abscissa attribute
(character)}

\item{\code{ord}}{The ordinate attribute
(character)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$reset(abj, abs = "time", ord = "infected")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-abscissaIsNumeric"></a>}}
\if{latex}{\out{\hypertarget{method-abscissaIsNumeric}{}}}
\subsection{Method \code{abscissaIsNumeric()}}{
Tests if the abscissa attribute is of type numeric.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.explorer$abscissaIsNumeric()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Test result
(logical)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{y <- x$abscissaIsNumeric()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ordinateIsNumeric"></a>}}
\if{latex}{\out{\hypertarget{method-ordinateIsNumeric}{}}}
\subsection{Method \code{ordinateIsNumeric()}}{
Tests if the ordinate attribute is of type numeric.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.explorer$ordinateIsNumeric()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Test result
(logical)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{y <- x$ordinateIsNumeric()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scatterPlot"></a>}}
\if{latex}{\out{\hypertarget{method-scatterPlot}{}}}
\subsection{Method \code{scatterPlot()}}{
Creates and returns a scatter plot abscissa and ordinate value pairs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.explorer$scatterPlot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Scatter plot
(ggplot2::ggplot)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$scatterPlot() \%>\%
 show()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-abscissaBarPlot"></a>}}
\if{latex}{\out{\hypertarget{method-abscissaBarPlot}{}}}
\subsection{Method \code{abscissaBarPlot()}}{
Creates and returns a histogram from the abscissa values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.explorer$abscissaBarPlot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Bar plot
(ggplot2::ggplot)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$abscissaBarPlot() \%>\%
 show()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-abscissaBoxPlot"></a>}}
\if{latex}{\out{\hypertarget{method-abscissaBoxPlot}{}}}
\subsection{Method \code{abscissaBoxPlot()}}{
Creates and returns a box plot from the abscissa values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.explorer$abscissaBoxPlot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Box plot
(ggplot2::ggplot)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$abscissaBoxPlot() \%>\%
 show()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-abscissaPlot"></a>}}
\if{latex}{\out{\hypertarget{method-abscissaPlot}{}}}
\subsection{Method \code{abscissaPlot()}}{
Creates and returns a compund graphical analysis of the abscissa values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.explorer$abscissaPlot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Compound plot
(gridExtra::grid.arrange)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$abscissaPlot() \%>\%
 show()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ordinateBarPlot"></a>}}
\if{latex}{\out{\hypertarget{method-ordinateBarPlot}{}}}
\subsection{Method \code{ordinateBarPlot()}}{
Creates and returns a histogram from the ordinate values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.explorer$ordinateBarPlot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Bar plot
(ggplot2::ggplot)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$ordinateBarPlot() \%>\%
 show()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ordinateBoxPlot"></a>}}
\if{latex}{\out{\hypertarget{method-ordinateBoxPlot}{}}}
\subsection{Method \code{ordinateBoxPlot()}}{
Creates and returns a box plot from the ordinate values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.explorer$ordinateBoxPlot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Box plot
(ggplot2::ggplot)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$ordinateBoxPlot() \%>\%
 show()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ordinatePlot"></a>}}
\if{latex}{\out{\hypertarget{method-ordinatePlot}{}}}
\subsection{Method \code{ordinatePlot()}}{
Creates and returns a compund graphical analysis of the ordinate values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.explorer$ordinatePlot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Compound plot
(gridExtra::grid.arrange)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$ordinatePlot() \%>\%
 show()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarizeNumeric"></a>}}
\if{latex}{\out{\hypertarget{method-summarizeNumeric}{}}}
\subsection{Method \code{summarizeNumeric()}}{
Summarizes the numeric values of a vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.explorer$summarizeNumeric(val = "numeric")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{values to analyze
(numeric)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
summary
(tibble::tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$summarizeNumeric(val) \%>\%
 print()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-abscissaStatistic"></a>}}
\if{latex}{\out{\hypertarget{method-abscissaStatistic}{}}}
\subsection{Method \code{abscissaStatistic()}}{
Calculates and returns the statistics of the abscissa values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.explorer$abscissaStatistic()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
statistics
(tibble::tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$abscissaStatistic() \%>\%
 print()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ordinateStatistic"></a>}}
\if{latex}{\out{\hypertarget{method-ordinateStatistic}{}}}
\subsection{Method \code{ordinateStatistic()}}{
Calculates and returns the statistics of the ordinate values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.explorer$ordinateStatistic()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
statistics
(tibble::tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$ordinateStatistic() \%>\%
 print()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.explorer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
