% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pguImputation.R
\name{pgu.imputation}
\alias{pgu.imputation}
\title{pgu.imputation}
\format{
\link[R6:R6Class]{R6::R6Class} object.
}
\description{
Detects and substitutes missing values from data set.
}
\details{
Detects missing values in the normalized and Z-score transfromed data set.
Replace missing values with substitutes by classical and AI-powerd substitution algorithms.
}
\section{Construction}{

x <- pguIMP::pgu.imputation$new(data)
}

\examples{

## ------------------------------------------------
## Method `pgu.imputation$new`
## ------------------------------------------------

y <- tibble:tibble()
x <- pguIMP:pgu.imputation$new(data = y)

## ------------------------------------------------
## Method `pgu.imputation$print`
## ------------------------------------------------

x$print()
print(x)

## ------------------------------------------------
## Method `pgu.imputation$resetImputationParameter`
## ------------------------------------------------

x$resetImputtaionParameter(data)

## ------------------------------------------------
## Method `pgu.imputation$featureIdx`
## ------------------------------------------------

idx <- x$featureIdx(feature = "infected")

## ------------------------------------------------
## Method `pgu.imputation$filterFeatures`
## ------------------------------------------------

idx <- x$filterFeatures(data)

## ------------------------------------------------
## Method `pgu.imputation$imputationSiteIdxByFeature`
## ------------------------------------------------

idx <- x$imputationIdxByFeature(featureName = "infected")

## ------------------------------------------------
## Method `pgu.imputation$nanFeatureList`
## ------------------------------------------------

idx <- x$nanFeatureList(data)

## ------------------------------------------------
## Method `pgu.imputation$gatherMeasurements`
## ------------------------------------------------

idx <- x$gatherMeasurements(value)

## ------------------------------------------------
## Method `pgu.imputation$gatherMissings`
## ------------------------------------------------

idx <- x$gatherMissings(value)

## ------------------------------------------------
## Method `pgu.imputation$gatherExistings`
## ------------------------------------------------

idx <- x$gatherExistings(value)

## ------------------------------------------------
## Method `pgu.imputation$gatherFractionOfMissings`
## ------------------------------------------------

idx <- x$gatherFractionOfMissings(value)

## ------------------------------------------------
## Method `pgu.imputation$gatherImputationStatistics`
## ------------------------------------------------

x$gatherInformationStatistics(value)

## ------------------------------------------------
## Method `pgu.imputation$detectImputationSites`
## ------------------------------------------------

y <- x$detectImputationsites(data, feature = "infected")

## ------------------------------------------------
## Method `pgu.imputation$handleImputationSites`
## ------------------------------------------------

x$handleImputationSites(data, progress)

## ------------------------------------------------
## Method `pgu.imputation$imputeByMedian`
## ------------------------------------------------

x$imputeByMedian(data, progress)

## ------------------------------------------------
## Method `pgu.imputation$imputeByMean`
## ------------------------------------------------

x$imputeByMean(data, progress)

## ------------------------------------------------
## Method `pgu.imputation$imputeByExpectationValue`
## ------------------------------------------------

x$imputeByExpectationValue(data, progress)

## ------------------------------------------------
## Method `pgu.imputation$imputeByMC`
## ------------------------------------------------

x$imputeByMC(data, progress)

## ------------------------------------------------
## Method `pgu.imputation$imputeByKnn`
## ------------------------------------------------

x$imputeByKnn(data, progress)

## ------------------------------------------------
## Method `pgu.imputation$imputeByMice`
## ------------------------------------------------

x$imputeByMice(data, progress)

## ------------------------------------------------
## Method `pgu.imputation$imputeByM5P`
## ------------------------------------------------

x$imputeByM5P(data, progress)

## ------------------------------------------------
## Method `pgu.imputation$imputeByAmelia`
## ------------------------------------------------

x$imputeByAmelia(data, progress)

## ------------------------------------------------
## Method `pgu.imputation$imputeByAmeliaBound`
## ------------------------------------------------

x$imputeByAmeliaBound(data, progress)

## ------------------------------------------------
## Method `pgu.imputation$imputationSiteDistribution`
## ------------------------------------------------

data \%>\%
 x$imputationSiteDistribution() \%>\%
 print()

## ------------------------------------------------
## Method `pgu.imputation$mergeImputationSiteData`
## ------------------------------------------------

data \%>\%
 x$mergeImputationSiteData() \%>\%
 print()

## ------------------------------------------------
## Method `pgu.imputation$imputationSiteHeatMap`
## ------------------------------------------------

x$imputationSiteHeatMap() \%>\%
 show()

## ------------------------------------------------
## Method `pgu.imputation$featureBarPlot`
## ------------------------------------------------

x$featureBarPlot() \%>\%
 show()

## ------------------------------------------------
## Method `pgu.imputation$featureBoxPlotWithSubset`
## ------------------------------------------------

x$featureBoxPlotWithSubset() \%>\%
 show()

## ------------------------------------------------
## Method `pgu.imputation$featurePlot`
## ------------------------------------------------

x$featurePlot() \%>\%
 show()
}
\author{
Sebastian Malkusch, \email{malkusch@med.uni-frankfurt.de}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{imputationParameter}}{Returns the instance variable outliersParameter.
(tibble::tibble)}

\item{\code{imputationSites}}{Returns the instance variable imputationSites.
(tibble::tibble)}

\item{\code{imputationAgentAlphabet}}{Returns the instance variable imputationagentAlphabet.
(character)}

\item{\code{imputationAgent}}{Returns the instance variable imputationAgent.
(character)}

\item{\code{setImputationAgent}}{Sets the instance variable imputationAgent.
(character)}

\item{\code{seed}}{Returns the instance variable seed.
(numeric)}

\item{\code{setSeed}}{Sets the instance variable seed.
(numeric)}

\item{\code{iterations}}{Returns the instance variable iterations.
(numeric)}

\item{\code{setIterations}}{Sets the instance variable iterations.
(numeric)}

\item{\code{amv}}{Returns the instance variable amv.
(numeric)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{pgu.imputation$new()}}
\item \href{#method-finalize}{\code{pgu.imputation$finalize()}}
\item \href{#method-print}{\code{pgu.imputation$print()}}
\item \href{#method-resetImputationParameter}{\code{pgu.imputation$resetImputationParameter()}}
\item \href{#method-featureIdx}{\code{pgu.imputation$featureIdx()}}
\item \href{#method-filterFeatures}{\code{pgu.imputation$filterFeatures()}}
\item \href{#method-imputationSiteIdxByFeature}{\code{pgu.imputation$imputationSiteIdxByFeature()}}
\item \href{#method-nanFeatureList}{\code{pgu.imputation$nanFeatureList()}}
\item \href{#method-gatherMeasurements}{\code{pgu.imputation$gatherMeasurements()}}
\item \href{#method-gatherMissings}{\code{pgu.imputation$gatherMissings()}}
\item \href{#method-gatherExistings}{\code{pgu.imputation$gatherExistings()}}
\item \href{#method-gatherFractionOfMissings}{\code{pgu.imputation$gatherFractionOfMissings()}}
\item \href{#method-gatherImputationStatistics}{\code{pgu.imputation$gatherImputationStatistics()}}
\item \href{#method-detectImputationSites}{\code{pgu.imputation$detectImputationSites()}}
\item \href{#method-handleImputationSites}{\code{pgu.imputation$handleImputationSites()}}
\item \href{#method-imputeByMedian}{\code{pgu.imputation$imputeByMedian()}}
\item \href{#method-imputeByMean}{\code{pgu.imputation$imputeByMean()}}
\item \href{#method-imputeByExpectationValue}{\code{pgu.imputation$imputeByExpectationValue()}}
\item \href{#method-imputeByMC}{\code{pgu.imputation$imputeByMC()}}
\item \href{#method-imputeByKnn}{\code{pgu.imputation$imputeByKnn()}}
\item \href{#method-imputeByMice}{\code{pgu.imputation$imputeByMice()}}
\item \href{#method-imputeByM5P}{\code{pgu.imputation$imputeByM5P()}}
\item \href{#method-imputeByAmelia}{\code{pgu.imputation$imputeByAmelia()}}
\item \href{#method-imputeByAmeliaBound}{\code{pgu.imputation$imputeByAmeliaBound()}}
\item \href{#method-imputationSiteDistribution}{\code{pgu.imputation$imputationSiteDistribution()}}
\item \href{#method-mergeImputationSiteData}{\code{pgu.imputation$mergeImputationSiteData()}}
\item \href{#method-imputationSiteHeatMap}{\code{pgu.imputation$imputationSiteHeatMap()}}
\item \href{#method-featureBarPlot}{\code{pgu.imputation$featureBarPlot()}}
\item \href{#method-featureBoxPlotWithSubset}{\code{pgu.imputation$featureBoxPlotWithSubset()}}
\item \href{#method-featurePlot}{\code{pgu.imputation$featurePlot()}}
\item \href{#method-clone}{\code{pgu.imputation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates and returns a new \code{pgu.imputation} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$new(data = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data to be cleaned.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{pgu.imputation} object.
(pguIMP::pgu.imputation)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{y <- tibble:tibble()
x <- pguIMP:pgu.imputation$new(data = y)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Clears the heap and
indicates that instance of \code{pgu.imputation} is removed from heap.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints instance variables of a \code{pgu.imputation} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
string
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$print()
print(x)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-resetImputationParameter"></a>}}
\if{latex}{\out{\hypertarget{method-resetImputationParameter}{}}}
\subsection{Method \code{resetImputationParameter()}}{
Resets instance variables and
identifies missings in the normalized and Z-score transfromed data set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$resetImputationParameter(data = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Dataframe to be analyzed.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$resetImputtaionParameter(data)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featureIdx"></a>}}
\if{latex}{\out{\hypertarget{method-featureIdx}{}}}
\subsection{Method \code{featureIdx()}}{
Returns the position of an attribute within a data frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$featureIdx(feature = "character")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature}}{The attribute's name.
(character)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The postion of the attribute.
(numeric)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{idx <- x$featureIdx(feature = "infected")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filterFeatures"></a>}}
\if{latex}{\out{\hypertarget{method-filterFeatures}{}}}
\subsection{Method \code{filterFeatures()}}{
Selects features cotaining missing values from a dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$filterFeatures(data = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Dataframe to be analyzed.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The filtered data frame.
(tibble::tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{idx <- x$filterFeatures(data)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-imputationSiteIdxByFeature"></a>}}
\if{latex}{\out{\hypertarget{method-imputationSiteIdxByFeature}{}}}
\subsection{Method \code{imputationSiteIdxByFeature()}}{
Returns the position of an attribute's missing values within a data frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$imputationSiteIdxByFeature(featureName = "character")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{featureName}}{The attribute's name.
(character)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The postion of the missing values.
(numeric)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{idx <- x$imputationIdxByFeature(featureName = "infected")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nanFeatureList"></a>}}
\if{latex}{\out{\hypertarget{method-nanFeatureList}{}}}
\subsection{Method \code{nanFeatureList()}}{
Characterizes each row of the data frame as either \code{complete}
or indicates which attribute are missing within the row.
If multiple attributes' row entries are missing, the row is characterized by \code{multiple}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$nanFeatureList(data = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data frame to be analyzed.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of row characteristics.
(character)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{idx <- x$nanFeatureList(data)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gatherMeasurements"></a>}}
\if{latex}{\out{\hypertarget{method-gatherMeasurements}{}}}
\subsection{Method \code{gatherMeasurements()}}{
Calculates the number of values of a vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$gatherMeasurements(value = "numeric")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A vector comprising numeric data.
(numeric)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The lenght of the vector.
(numeric)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{idx <- x$gatherMeasurements(value)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gatherMissings"></a>}}
\if{latex}{\out{\hypertarget{method-gatherMissings}{}}}
\subsection{Method \code{gatherMissings()}}{
Calculates the number of missing values of a vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$gatherMissings(value = "numeric")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A vector comprising numeric data.
(numeric)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The number of missing in the vector.
(numeric)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{idx <- x$gatherMissings(value)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gatherExistings"></a>}}
\if{latex}{\out{\hypertarget{method-gatherExistings}{}}}
\subsection{Method \code{gatherExistings()}}{
Calculates the number of existing values of a vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$gatherExistings(value = "numeric")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A vector comprising numeric data.
(numeric)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The number of existing values in the vector.
(numeric)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{idx <- x$gatherExistings(value)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gatherFractionOfMissings"></a>}}
\if{latex}{\out{\hypertarget{method-gatherFractionOfMissings}{}}}
\subsection{Method \code{gatherFractionOfMissings()}}{
Calculates the fraction of missing values of a vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$gatherFractionOfMissings(value = "numeric")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A vector comprising numeric data.
(numeric)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The fraction of missing values in the vector.
(numeric)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{idx <- x$gatherFractionOfMissings(value)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gatherImputationStatistics"></a>}}
\if{latex}{\out{\hypertarget{method-gatherImputationStatistics}{}}}
\subsection{Method \code{gatherImputationStatistics()}}{
Gathers statistical information about missing values
that are provided by the classes public \code{gather} functions.
The information is stored within the classes instance variable \code{imputationParameter}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$gatherImputationStatistics(data = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data frame to be analyzed.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$gatherInformationStatistics(value)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-detectImputationSites"></a>}}
\if{latex}{\out{\hypertarget{method-detectImputationSites}{}}}
\subsection{Method \code{detectImputationSites()}}{
Detects missing values within the data frame and
writes the to the instance variable \code{imputationsites}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$detectImputationSites(data = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data frame to be analyzed.
(tibble::tibble)}

\item{\code{feature}}{The attribute within the data frame to be analyzed.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{y <- x$detectImputationsites(data, feature = "infected")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-handleImputationSites"></a>}}
\if{latex}{\out{\hypertarget{method-handleImputationSites}{}}}
\subsection{Method \code{handleImputationSites()}}{
Chooses a cleaning method based upon the instance variable \code{imputationAgent}
and handles the imputation sites in the dataframe.
Returns a cleaned data set.
Display the progress if shiny is loaded.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$handleImputationSites(data = "tbl_df", progress = "Progress")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data frame to be analyzed.
(tibble::tibble)}

\item{\code{progress}}{If shiny is loaded, the analysis' progress is stored within this instance of the shiny Progress class.
(shiny::Progress)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Cleaned dataframe.
(tibble:tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$handleImputationSites(data, progress)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-imputeByMedian"></a>}}
\if{latex}{\out{\hypertarget{method-imputeByMedian}{}}}
\subsection{Method \code{imputeByMedian()}}{
Substitutes imputation sites by the median of the respective attribute.
Returns the cleaned dataframe.
Display the progress if shiny is loaded.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$imputeByMedian(data = "tbl_df", progress = "Progress")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data frame to be analyzed.
(tibble::tibble)}

\item{\code{progress}}{If shiny is loaded, the analysis' progress is stored in this instance of the shiny Progress class.
(shiny::Progress)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Cleaned dataframe.
(tibble:tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$imputeByMedian(data, progress)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-imputeByMean"></a>}}
\if{latex}{\out{\hypertarget{method-imputeByMean}{}}}
\subsection{Method \code{imputeByMean()}}{
Substitutes imputation sites by the aritmertic mean of the respective attribute.
Returns the cleaned dataframe.
Display the progress if shiny is loaded.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$imputeByMean(data = "tbl_df", progress = "Progress")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data frame to be analyzed.
(tibble::tibble)}

\item{\code{progress}}{If shiny is loaded, the analysis' progress is stored in this instance of the shiny Progress class.
(shiny::Progress)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Cleaned dataframe.
(tibble:tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$imputeByMean(data, progress)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-imputeByExpectationValue"></a>}}
\if{latex}{\out{\hypertarget{method-imputeByExpectationValue}{}}}
\subsection{Method \code{imputeByExpectationValue()}}{
Substitutes imputation sites by the expectation value of the respective attribute.
Returns the cleaned dataframe.
Display the progress if shiny is loaded.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$imputeByExpectationValue(data = "tbl_df", progress = "Progress")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data frame to be analyzed.
(tibble::tibble)}

\item{\code{progress}}{If shiny is loaded, the analysis' progress is stored in this instance of the shiny Progress class.
(shiny::Progress)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Cleaned dataframe.
(tibble:tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$imputeByExpectationValue(data, progress)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-imputeByMC"></a>}}
\if{latex}{\out{\hypertarget{method-imputeByMC}{}}}
\subsection{Method \code{imputeByMC()}}{
Substitutes imputation sites by values generated by a monte carlo simulation.
The procedure runs several times as defined by the instance variable \code{iterations}.
The run with the best result is identified and used for substitution.
Returns the cleaned dataframe.
Display the progress if shiny is loaded.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$imputeByMC(data = "tbl_df", progress = "Progress")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data frame to be analyzed.
(tibble::tibble)}

\item{\code{progress}}{If shiny is loaded, the analysis' progress is stored in this instance of the shiny Progress class.
(shiny::Progress)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Cleaned dataframe.
(tibble:tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$imputeByMC(data, progress)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-imputeByKnn"></a>}}
\if{latex}{\out{\hypertarget{method-imputeByKnn}{}}}
\subsection{Method \code{imputeByKnn()}}{
Substitutes imputation sites by predictions of a KNN analysis of the whole dataframe.
Returns the cleaned dataframe.
Display the progress if shiny is loaded.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$imputeByKnn(data = "tbl_df", progress = "Progress")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data frame to be analyzed.
(tibble::tibble)}

\item{\code{progress}}{If shiny is loaded, the analysis' progress is stored in this instance of the shiny Progress class.
(shiny::Progress)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Cleaned dataframe.
(tibble:tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$imputeByKnn(data, progress)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-imputeByMice"></a>}}
\if{latex}{\out{\hypertarget{method-imputeByMice}{}}}
\subsection{Method \code{imputeByMice()}}{
Substitutes imputation sites by values generated by a different methods of the mice package.
The procedure runs several times as defined by the instance variable \code{iterations}.
The run with the best result is identified and used for substitution.
Returns the cleaned dataframe.
Display the progress if shiny is loaded.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$imputeByMice(data, method = "character", progress = "Progress")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data frame to be analyzed.
(tibble::tibble)}

\item{\code{method}}{One of the methods supported by the mice package.
(character)}

\item{\code{progress}}{If shiny is loaded, the analysis' progress is stored in this instance of the shiny Progress class.
(shiny::Progress)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Cleaned dataframe.
(tibble:tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$imputeByMice(data, progress)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-imputeByM5P"></a>}}
\if{latex}{\out{\hypertarget{method-imputeByM5P}{}}}
\subsection{Method \code{imputeByM5P()}}{
Substitutes imputation sites by predictions of a M5P tree trained on the whole dataframe.
Returns the cleaned dataframe.
Display the progress if shiny is loaded.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$imputeByM5P(data = "tl_df", progress = "Progress")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data frame to be analyzed.
(tibble::tibble)}

\item{\code{progress}}{If shiny is loaded, the analysis' progress is stored in this instance of the shiny Progress class.
(shiny::Progress)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Cleaned dataframe.
(tibble:tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$imputeByM5P(data, progress)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-imputeByAmelia"></a>}}
\if{latex}{\out{\hypertarget{method-imputeByAmelia}{}}}
\subsection{Method \code{imputeByAmelia()}}{
Substitutes imputation sites by values generated by a different methods of the amelia package.
The procedure is run without bound information.
Returns the cleaned dataframe.
Display the progress if shiny is loaded.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$imputeByAmelia(data = "tbl_df", progress = "Progress")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data frame to be analyzed.
(tibble::tibble)}

\item{\code{progress}}{If shiny is loaded, the analysis' progress is stored in this instance of the shiny Progress class.
(shiny::Progress)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Cleaned dataframe.
(tibble:tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$imputeByAmelia(data, progress)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-imputeByAmeliaBound"></a>}}
\if{latex}{\out{\hypertarget{method-imputeByAmeliaBound}{}}}
\subsection{Method \code{imputeByAmeliaBound()}}{
Substitutes imputation sites by values generated by a different methods of the amelia package.
The procedure is run with bound information.
Returns the cleaned dataframe.
Display the progress if shiny is loaded.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$imputeByAmeliaBound(data = "tbl_df", progress = "Progress")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data frame to be analyzed.
(tibble::tibble)}

\item{\code{progress}}{If shiny is loaded, the analysis' progress is stored in this instance of the shiny Progress class.
(shiny::Progress)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Cleaned dataframe.
(tibble:tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$imputeByAmeliaBound(data, progress)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-imputationSiteDistribution"></a>}}
\if{latex}{\out{\hypertarget{method-imputationSiteDistribution}{}}}
\subsection{Method \code{imputationSiteDistribution()}}{
Numeric representation of the distribution of missing values within the data frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$imputationSiteDistribution(data = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data frame to be analyzed.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame
(tibble::tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{data \%>\%
 x$imputationSiteDistribution() \%>\%
 print()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mergeImputationSiteData"></a>}}
\if{latex}{\out{\hypertarget{method-mergeImputationSiteData}{}}}
\subsection{Method \code{mergeImputationSiteData()}}{
Merges the numeric attributes of the pguIMP data with its metadata.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$mergeImputationSiteData(
  dfData = "tbl_df",
  dfMetadata = "tbl_df"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dfData}}{The data frame to be analyzed.
(tibble::tibble)}

\item{\code{dfMetadata}}{The corresponding metadata.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame
(tibble::tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{data \%>\%
 x$mergeImputationSiteData() \%>\%
 print()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-imputationSiteHeatMap"></a>}}
\if{latex}{\out{\hypertarget{method-imputationSiteHeatMap}{}}}
\subsection{Method \code{imputationSiteHeatMap()}}{
Displays the distribution of missing values in form of a heatmap.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$imputationSiteHeatMap()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A heatmap plot.
(ggplot2::ggplot)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$imputationSiteHeatMap() \%>\%
 show()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featureBarPlot"></a>}}
\if{latex}{\out{\hypertarget{method-featureBarPlot}{}}}
\subsection{Method \code{featureBarPlot()}}{
Displays the distribution of an attribute values as histogram.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$featureBarPlot(data = "tbl_df", feature = "character")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{dataframe to be analyzed.
(tibble::tibble)}

\item{\code{feature}}{attribute to be shown.
(character)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A histogram.
(ggplot2::ggplot)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$featureBarPlot() \%>\%
 show()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featureBoxPlotWithSubset"></a>}}
\if{latex}{\out{\hypertarget{method-featureBoxPlotWithSubset}{}}}
\subsection{Method \code{featureBoxPlotWithSubset()}}{
Displays the distribution of an attribute's values as box plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$featureBoxPlotWithSubset(data = "tbl_df", feature = "character")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{dataframe to be analyzed.
(tibble::tibble)}

\item{\code{feature}}{attribute to be shown.
(character)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A box plot.
(ggplot2::ggplot)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$featureBoxPlotWithSubset() \%>\%
 show()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featurePlot"></a>}}
\if{latex}{\out{\hypertarget{method-featurePlot}{}}}
\subsection{Method \code{featurePlot()}}{
Displays the distribution of an attribute's values as a composition of a box plot and a histogram.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$featurePlot(data = "tbl_df", feature = "character")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{dataframe to be analyzed.
(tibble::tibble)}

\item{\code{feature}}{attribute to be shown.
(character)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A composite plot.
(ggplot2::ggplot)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$featurePlot() \%>\%
 show()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.imputation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
