% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pguLimitsOfQuantification.R
\name{pgu.limitsOfQuantification}
\alias{pgu.limitsOfQuantification}
\title{pgu.limitsOfQuantification}
\format{
\link[R6:R6Class]{R6::R6Class} object.
}
\description{
Handles values in the pguIMP dataset that exceed the limits of quantification.
}
\details{
more information
}
\section{Construction}{

x <- pguIMP::pgu.limitsOfQuantification$new(data)
}

\examples{

## ------------------------------------------------
## Method `pgu.limitsOfQuantification$new`
## ------------------------------------------------

y <- tibble:tibble()
x <- pguIMP:pgu.limitsOfQuantification$new(data = y)

## ------------------------------------------------
## Method `pgu.limitsOfQuantification$print`
## ------------------------------------------------

x$print()
print(x)

## ------------------------------------------------
## Method `pgu.limitsOfQuantification$checkValidity`
## ------------------------------------------------

y <- x$checkValidity(features)

## ------------------------------------------------
## Method `pgu.limitsOfQuantification$resetLoqStatistics`
## ------------------------------------------------

x$resetLoqStatistics(obj)

## ------------------------------------------------
## Method `pgu.limitsOfQuantification$resetLoq`
## ------------------------------------------------

x$resetLoq()

## ------------------------------------------------
## Method `pgu.limitsOfQuantification$featureLloq`
## ------------------------------------------------

lloq <- x$featureLloq(feature = "infected")

## ------------------------------------------------
## Method `pgu.limitsOfQuantification$featureUloq`
## ------------------------------------------------

uloq <- x$featureUloq(feature = "infected")

## ------------------------------------------------
## Method `pgu.limitsOfQuantification$appendOutlier`
## ------------------------------------------------

x$appendOutlier(feature = "infected", value = 501, idx = 200, lloq=0, uloq=100)

## ------------------------------------------------
## Method `pgu.limitsOfQuantification$findOutliers`
## ------------------------------------------------

y <- x$findOutliers(obj)

## ------------------------------------------------
## Method `pgu.limitsOfQuantification$collectStatistics`
## ------------------------------------------------

x$collectStatistics(obj)

## ------------------------------------------------
## Method `pgu.limitsOfQuantification$mutateLoqOutliers`
## ------------------------------------------------

y <- x$mutateLoqOutliers(obj)

## ------------------------------------------------
## Method `pgu.limitsOfQuantification$mutateLloqOutliers`
## ------------------------------------------------

y <- x$mutateLloqOutliers(obj)

## ------------------------------------------------
## Method `pgu.limitsOfQuantification$mutateUloqOutliers`
## ------------------------------------------------

y <- x$mutateUloqOutliers(obj)

## ------------------------------------------------
## Method `pgu.limitsOfQuantification$featureOutlier`
## ------------------------------------------------

f <- x$featureOutlier(feature = "infected")

## ------------------------------------------------
## Method `pgu.limitsOfQuantification$dataInformation`
## ------------------------------------------------

x$dataInformation() \%>\%
 print()

## ------------------------------------------------
## Method `pgu.limitsOfQuantification$loqDataTable`
## ------------------------------------------------

x$loqDataTable(dfData, dfMetadata) \%>\%
 show()

## ------------------------------------------------
## Method `pgu.limitsOfQuantification$loqFeatureTable`
## ------------------------------------------------

x$loqFeatureTable(obj, feature = "infected") \%>\%
 show()

## ------------------------------------------------
## Method `pgu.limitsOfQuantification$plotLoqDistribution`
## ------------------------------------------------

x$plotLoqDistribution() \%>\%
 show()

## ------------------------------------------------
## Method `pgu.limitsOfQuantification$featureBarPlot`
## ------------------------------------------------

x$featureBarPlot(obj, feature = "infected") \%>\%
 show()

## ------------------------------------------------
## Method `pgu.limitsOfQuantification$featureBoxPlotWithSubset`
## ------------------------------------------------

x$featureBoxPlotWithSubset(obj, feature = "infected") \%>\%
 show()

## ------------------------------------------------
## Method `pgu.limitsOfQuantification$featurePlot`
## ------------------------------------------------

x$featurePlot(obj, feature = "infected") \%>\%
 show()
}
\author{
Sebastian Malkusch, \email{malkusch@med.uni-frankfurt.de}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{loq}}{Returns the instance variable loq
(tibble::tibble)}

\item{\code{setLoq}}{Sets the instance variable loq.
(tibble::tibble)}

\item{\code{outliers}}{Returns instance variable outliers
(tibble::tibble)}

\item{\code{lloqSubstituteAlphabet}}{Returns the instance variable lloqSubstititeAlphabet
(character)}

\item{\code{lloqSubstituteAgent}}{Returns the instance variable lloqSubstituteAgent
(character)}

\item{\code{setLloqSubstituteAgent}}{Sets the instance variable lloqSubstituteAgent.
(character)}

\item{\code{uloqSubstituteAlphabet}}{Returns the instance variable uloqSubstititeAlphabet
(character)}

\item{\code{uloqSubstituteAgent}}{Returns the instance variable uloqSubstituteAgent
(character)}

\item{\code{setUloqSubstituteAgent}}{Sets the instance variable uloqSubstituteAgent.
(character)}

\item{\code{naHandlingAlphabet}}{Returns the instance variable naHandlingAlphabet
(character)}

\item{\code{naHandlingAgent}}{Returns the instance variable naHandlingAgentt
(character)}

\item{\code{setNaHandlingAgent}}{Sets the instance variable naHandlingAgentt
(character)}

\item{\code{loqStatistics}}{Returns the instance variable loqStatistics}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{pgu.limitsOfQuantification$new()}}
\item \href{#method-finalize}{\code{pgu.limitsOfQuantification$finalize()}}
\item \href{#method-print}{\code{pgu.limitsOfQuantification$print()}}
\item \href{#method-checkValidity}{\code{pgu.limitsOfQuantification$checkValidity()}}
\item \href{#method-resetLoqStatistics}{\code{pgu.limitsOfQuantification$resetLoqStatistics()}}
\item \href{#method-resetLoq}{\code{pgu.limitsOfQuantification$resetLoq()}}
\item \href{#method-featureLloq}{\code{pgu.limitsOfQuantification$featureLloq()}}
\item \href{#method-featureUloq}{\code{pgu.limitsOfQuantification$featureUloq()}}
\item \href{#method-appendOutlier}{\code{pgu.limitsOfQuantification$appendOutlier()}}
\item \href{#method-findOutliers}{\code{pgu.limitsOfQuantification$findOutliers()}}
\item \href{#method-collectStatistics}{\code{pgu.limitsOfQuantification$collectStatistics()}}
\item \href{#method-mutateLoqOutliers}{\code{pgu.limitsOfQuantification$mutateLoqOutliers()}}
\item \href{#method-mutateLloqOutliers}{\code{pgu.limitsOfQuantification$mutateLloqOutliers()}}
\item \href{#method-mutateUloqOutliers}{\code{pgu.limitsOfQuantification$mutateUloqOutliers()}}
\item \href{#method-featureOutlier}{\code{pgu.limitsOfQuantification$featureOutlier()}}
\item \href{#method-dataInformation}{\code{pgu.limitsOfQuantification$dataInformation()}}
\item \href{#method-loqDataTable}{\code{pgu.limitsOfQuantification$loqDataTable()}}
\item \href{#method-loqFeatureTable}{\code{pgu.limitsOfQuantification$loqFeatureTable()}}
\item \href{#method-plotLoqDistribution}{\code{pgu.limitsOfQuantification$plotLoqDistribution()}}
\item \href{#method-featureBarPlot}{\code{pgu.limitsOfQuantification$featureBarPlot()}}
\item \href{#method-featureBoxPlotWithSubset}{\code{pgu.limitsOfQuantification$featureBoxPlotWithSubset()}}
\item \href{#method-featurePlot}{\code{pgu.limitsOfQuantification$featurePlot()}}
\item \href{#method-clone}{\code{pgu.limitsOfQuantification$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates and returns a new \code{pgu.limitsOfQuantification} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.limitsOfQuantification$new(obj = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{pgu.limitsOfQuantification} object.
(pguIMP::pgu.optimizer)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{y <- tibble:tibble()
x <- pguIMP:pgu.limitsOfQuantification$new(data = y)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Clears the heap and
indicates that instance of \code{pgu.limitaOfQuantification} is removed from heap.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.limitsOfQuantification$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints instance variables of a \code{pgu.limitsOfQuantification} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.limitsOfQuantification$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
string
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$print()
print(x)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkValidity"></a>}}
\if{latex}{\out{\hypertarget{method-checkValidity}{}}}
\subsection{Method \code{checkValidity()}}{
Tests if the provided attributes are known to the class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.limitsOfQuantification$checkValidity(features = "charater")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{features}}{A vector of attributes to be testes
(character)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Test result
(logical)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{y <- x$checkValidity(features)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-resetLoqStatistics"></a>}}
\if{latex}{\out{\hypertarget{method-resetLoqStatistics}{}}}
\subsection{Method \code{resetLoqStatistics()}}{
Resets the class' instance variable loqStatistics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.limitsOfQuantification$resetLoqStatistics(obj = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$resetLoqStatistics(obj)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-resetLoq"></a>}}
\if{latex}{\out{\hypertarget{method-resetLoq}{}}}
\subsection{Method \code{resetLoq()}}{
Resets the class' instance variable outliers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.limitsOfQuantification$resetLoq()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$resetLoq()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featureLloq"></a>}}
\if{latex}{\out{\hypertarget{method-featureLloq}{}}}
\subsection{Method \code{featureLloq()}}{
Returns the attribute's specific lloq.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.limitsOfQuantification$featureLloq(feature = "character")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature}}{The attribute to be analyzed
(character)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The attribute's lloq
(numeric)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{lloq <- x$featureLloq(feature = "infected")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featureUloq"></a>}}
\if{latex}{\out{\hypertarget{method-featureUloq}{}}}
\subsection{Method \code{featureUloq()}}{
Returns the attribute's specific uloq.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.limitsOfQuantification$featureUloq(feature = "character")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature}}{The attribute to be analyzed
(character)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The attribute's uloq
(numeric)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{uloq <- x$featureUloq(feature = "infected")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-appendOutlier"></a>}}
\if{latex}{\out{\hypertarget{method-appendOutlier}{}}}
\subsection{Method \code{appendOutlier()}}{
Extends the instance variable outliers by one entry.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.limitsOfQuantification$appendOutlier(
  feature = "character",
  value = "numeric",
  idx = "numeric",
  lloq = "numeric",
  uloq = "numeric"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature}}{The attribute the value is assigned to.
(character)}

\item{\code{value}}{The value to append
(numeric)}

\item{\code{idx}}{The index of the measruement the value is originated from.
(numeric)}

\item{\code{lloq}}{The attributes lloq
(numeric)}

\item{\code{uloq}}{The attributes uloq
(numeric)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$appendOutlier(feature = "infected", value = 501, idx = 200, lloq=0, uloq=100)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findOutliers"></a>}}
\if{latex}{\out{\hypertarget{method-findOutliers}{}}}
\subsection{Method \code{findOutliers()}}{
Searches for outliers in the given data frame.
If an outlier was found, it is appended to the instance variable outliers.
Indicates if an outlier was found.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.limitsOfQuantification$findOutliers(obj = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Test result
(logical)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{y <- x$findOutliers(obj)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-collectStatistics"></a>}}
\if{latex}{\out{\hypertarget{method-collectStatistics}{}}}
\subsection{Method \code{collectStatistics()}}{
Calculates statistics of outlier appearance.
Stores it into the instance variable loqStatistics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.limitsOfQuantification$collectStatistics(obj = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$collectStatistics(obj)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mutateLoqOutliers"></a>}}
\if{latex}{\out{\hypertarget{method-mutateLoqOutliers}{}}}
\subsection{Method \code{mutateLoqOutliers()}}{
Mutates all outlier candidates based on user defined actions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.limitsOfQuantification$mutateLoqOutliers(obj = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Te revised data frame
(tibble::tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{y <- x$mutateLoqOutliers(obj)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mutateLloqOutliers"></a>}}
\if{latex}{\out{\hypertarget{method-mutateLloqOutliers}{}}}
\subsection{Method \code{mutateLloqOutliers()}}{
Mutates outlier candidates characterized as below LLOQ based on user defined actions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.limitsOfQuantification$mutateLloqOutliers(obj = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Te revised data frame
(tibble::tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{y <- x$mutateLloqOutliers(obj)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mutateUloqOutliers"></a>}}
\if{latex}{\out{\hypertarget{method-mutateUloqOutliers}{}}}
\subsection{Method \code{mutateUloqOutliers()}}{
Mutates outlier candidates characterized as above ULOQ based on user defined actions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.limitsOfQuantification$mutateUloqOutliers(obj = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Te revised data frame
(tibble::tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{y <- x$mutateUloqOutliers(obj)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featureOutlier"></a>}}
\if{latex}{\out{\hypertarget{method-featureOutlier}{}}}
\subsection{Method \code{featureOutlier()}}{
Returns the detected outliers of a given attribute.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.limitsOfQuantification$featureOutlier(feature = "character")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature}}{The attribute to be analyzed
(character)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The attribute's outliers
(tibble::tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{f <- x$featureOutlier(feature = "infected")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataInformation"></a>}}
\if{latex}{\out{\hypertarget{method-dataInformation}{}}}
\subsection{Method \code{dataInformation()}}{
Gathers and returns class information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.limitsOfQuantification$dataInformation()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$dataInformation() \%>\%
 print()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loqDataTable"></a>}}
\if{latex}{\out{\hypertarget{method-loqDataTable}{}}}
\subsection{Method \code{loqDataTable()}}{
Merges dfData and dfMetadata and returns a fromatted data table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.limitsOfQuantification$loqDataTable(
  dfData = "tbl_df",
  dfMetadata = "tbl_df"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dfData}}{The data to be analyzed.
(tibble::tibble)}

\item{\code{dfMetadata}}{The data frame containing metadata.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A formatted data table
(DT::datatable)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$loqDataTable(dfData, dfMetadata) \%>\%
 show()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loqFeatureTable"></a>}}
\if{latex}{\out{\hypertarget{method-loqFeatureTable}{}}}
\subsection{Method \code{loqFeatureTable()}}{
Returns a formatted data table with comrising the information of a user defined attribute's outliers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.limitsOfQuantification$loqFeatureTable(
  obj = "tbl_df",
  feature = "character"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
(tibble::tibble)}

\item{\code{feature}}{The attribute to be analyzed
(character)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A formatted data table
(DT::datatable)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$loqFeatureTable(obj, feature = "infected") \%>\%
 show()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotLoqDistribution"></a>}}
\if{latex}{\out{\hypertarget{method-plotLoqDistribution}{}}}
\subsection{Method \code{plotLoqDistribution()}}{
Creates a plot of the instance variable loqStatistics.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.limitsOfQuantification$plotLoqDistribution()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A plot.
(ggplot2::ggplot)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$plotLoqDistribution() \%>\%
 show()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featureBarPlot"></a>}}
\if{latex}{\out{\hypertarget{method-featureBarPlot}{}}}
\subsection{Method \code{featureBarPlot()}}{
Creates a bar plot of a user defined attribute's value distribution.
LOQs are indicated as dotted lines
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.limitsOfQuantification$featureBarPlot(
  obj = "tbl_df",
  feature = "character"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
(tibble::tibble)}

\item{\code{feature}}{The attribute to be analyzed
(character)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A bar plot.
(ggplot2::ggplot)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$featureBarPlot(obj, feature = "infected") \%>\%
 show()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featureBoxPlotWithSubset"></a>}}
\if{latex}{\out{\hypertarget{method-featureBoxPlotWithSubset}{}}}
\subsection{Method \code{featureBoxPlotWithSubset()}}{
Creates a box plot of a user defined attribute's value distribution.
LOQs are indicated as dotted lines
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.limitsOfQuantification$featureBoxPlotWithSubset(
  obj = "tbl_df",
  feature = "character"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
(tibble::tibble)}

\item{\code{feature}}{The attribute to be analyzed
(character)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A box plot.
(ggplot2::ggplot)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$featureBoxPlotWithSubset(obj, feature = "infected") \%>\%
 show()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featurePlot"></a>}}
\if{latex}{\out{\hypertarget{method-featurePlot}{}}}
\subsection{Method \code{featurePlot()}}{
Creates and returns a composite graphical analysis
of the outlier analysis of a user defined attribute.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.limitsOfQuantification$featurePlot(obj = "tbl_df", feature = "character")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data to be analyzed.
(tibble::tibble)}

\item{\code{feature}}{Attribute's name.
(character)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Composite result plot.
(gridExtra::grid.arrange)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$featurePlot(obj, feature = "infected") \%>\%
 show()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.limitsOfQuantification$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
