% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pguMissings.R
\name{pgu.missings}
\alias{pgu.missings}
\title{pgu.missings}
\format{
\link[R6:R6Class]{R6::R6Class} object.
}
\description{
Detects and substitutes missing values from data set.
}
\details{
Detects missing values in the transformed and normalized data set.
}
\section{Construction}{

x <- pguIMP::pgu.missings$new(data)
}

\examples{

## ------------------------------------------------
## Method `pgu.missings$new`
## ------------------------------------------------

y <- tibble:tibble()
x <- pguIMP:pgu.missings$new(data_df = y)

## ------------------------------------------------
## Method `pgu.missings$print`
## ------------------------------------------------

x$print()
print(x)

## ------------------------------------------------
## Method `pgu.missings$resetImputationParameter`
## ------------------------------------------------

x$resetImputtaionParameter(data_df)

## ------------------------------------------------
## Method `pgu.missings$featureIdx`
## ------------------------------------------------

idx <- x$featureIdx(feature = "infected")

## ------------------------------------------------
## Method `pgu.missings$filterFeatures`
## ------------------------------------------------

idx <- x$filterFeatures(data_df)

## ------------------------------------------------
## Method `pgu.missings$gatherMeasurements`
## ------------------------------------------------

idx <- x$gatherMeasurements(value)

## ------------------------------------------------
## Method `pgu.missings$gatherMissings`
## ------------------------------------------------

idx <- x$gatherMissings(value)

## ------------------------------------------------
## Method `pgu.missings$gatherExistings`
## ------------------------------------------------

idx <- x$gatherExistings(value)

## ------------------------------------------------
## Method `pgu.missings$gatherFractionOfMissings`
## ------------------------------------------------

idx <- x$gatherFractionOfMissings(value)

## ------------------------------------------------
## Method `pgu.missings$gatherImputationStatistics`
## ------------------------------------------------

x$gatherInformationStatistics(data_df)

## ------------------------------------------------
## Method `pgu.missings$detectImputationSites`
## ------------------------------------------------

y <- x$detectImputationsites(data_df)

## ------------------------------------------------
## Method `pgu.missings$imputationSiteDistribution`
## ------------------------------------------------

data \%>\%
 x$imputationSiteDistribution() \%>\%
 print()

## ------------------------------------------------
## Method `pgu.missings$mergeImputationSiteData`
## ------------------------------------------------

data \%>\%
 x$mergeImputationSiteData() \%>\%
 print()

## ------------------------------------------------
## Method `pgu.missings$imputationSiteHeatMap`
## ------------------------------------------------

x$imputationSiteHeatMap() \%>\%
 show()
}
\author{
Sebastian Malkusch, \email{malkusch@med.uni-frankfurt.de}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{imputationParameter}}{Returns the instance variable outliersParameter.
(tibble::tibble)}

\item{\code{imputationSites}}{Returns the instance variable imputationSites.
(tibble::tibble)}

\item{\code{amv}}{Returns the instance variable amv.
(numeric)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{pgu.missings$new()}}
\item \href{#method-finalize}{\code{pgu.missings$finalize()}}
\item \href{#method-print}{\code{pgu.missings$print()}}
\item \href{#method-resetImputationParameter}{\code{pgu.missings$resetImputationParameter()}}
\item \href{#method-featureIdx}{\code{pgu.missings$featureIdx()}}
\item \href{#method-filterFeatures}{\code{pgu.missings$filterFeatures()}}
\item \href{#method-gatherMeasurements}{\code{pgu.missings$gatherMeasurements()}}
\item \href{#method-gatherMissings}{\code{pgu.missings$gatherMissings()}}
\item \href{#method-gatherExistings}{\code{pgu.missings$gatherExistings()}}
\item \href{#method-gatherFractionOfMissings}{\code{pgu.missings$gatherFractionOfMissings()}}
\item \href{#method-gatherImputationStatistics}{\code{pgu.missings$gatherImputationStatistics()}}
\item \href{#method-detectImputationSites}{\code{pgu.missings$detectImputationSites()}}
\item \href{#method-imputationSiteDistribution}{\code{pgu.missings$imputationSiteDistribution()}}
\item \href{#method-mergeImputationSiteData}{\code{pgu.missings$mergeImputationSiteData()}}
\item \href{#method-imputationSiteHeatMap}{\code{pgu.missings$imputationSiteHeatMap()}}
\item \href{#method-clone}{\code{pgu.missings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates and returns a new \code{pgu.missings} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.missings$new(data_df = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_df}}{The data to be cleaned.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{pgu.missings} object.
(pguIMP::pgu.missings)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{y <- tibble:tibble()
x <- pguIMP:pgu.missings$new(data_df = y)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Clears the heap and
indicates that instance of \code{pgu.missings} is removed from heap.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.missings$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints instance variables of a \code{pgu.missings} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.missings$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
string
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$print()
print(x)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-resetImputationParameter"></a>}}
\if{latex}{\out{\hypertarget{method-resetImputationParameter}{}}}
\subsection{Method \code{resetImputationParameter()}}{
Resets instance variables and
identifies missings in the normalized data set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.missings$resetImputationParameter(data_df = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_df}}{Dataframe to be analyzed.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$resetImputtaionParameter(data_df)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-featureIdx"></a>}}
\if{latex}{\out{\hypertarget{method-featureIdx}{}}}
\subsection{Method \code{featureIdx()}}{
Returns the position of an attribute within a data frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.missings$featureIdx(feature = "character")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature}}{The attribute's name.
(character)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The postion of the attribute.
(numeric)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{idx <- x$featureIdx(feature = "infected")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filterFeatures"></a>}}
\if{latex}{\out{\hypertarget{method-filterFeatures}{}}}
\subsection{Method \code{filterFeatures()}}{
Selects features cotaining missing values from a dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.missings$filterFeatures(data_df = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_df}}{Dataframe to be analyzed.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The filtered data frame.
(tibble::tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{idx <- x$filterFeatures(data_df)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gatherMeasurements"></a>}}
\if{latex}{\out{\hypertarget{method-gatherMeasurements}{}}}
\subsection{Method \code{gatherMeasurements()}}{
Calculates the number of values of a vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.missings$gatherMeasurements(value = "numeric")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A vector comprising numeric data.
(numeric)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The lenght of the vector.
(numeric)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{idx <- x$gatherMeasurements(value)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gatherMissings"></a>}}
\if{latex}{\out{\hypertarget{method-gatherMissings}{}}}
\subsection{Method \code{gatherMissings()}}{
Calculates the number of missing values of a vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.missings$gatherMissings(value = "numeric")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A vector comprising numeric data.
(numeric)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The number of missing in the vector.
(numeric)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{idx <- x$gatherMissings(value)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gatherExistings"></a>}}
\if{latex}{\out{\hypertarget{method-gatherExistings}{}}}
\subsection{Method \code{gatherExistings()}}{
Calculates the number of existing values of a vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.missings$gatherExistings(value = "numeric")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A vector comprising numeric data.
(numeric)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The number of existing values in the vector.
(numeric)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{idx <- x$gatherExistings(value)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gatherFractionOfMissings"></a>}}
\if{latex}{\out{\hypertarget{method-gatherFractionOfMissings}{}}}
\subsection{Method \code{gatherFractionOfMissings()}}{
Calculates the fraction of missing values of a vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.missings$gatherFractionOfMissings(value = "numeric")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A vector comprising numeric data.
(numeric)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The fraction of missing values in the vector.
(numeric)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{idx <- x$gatherFractionOfMissings(value)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gatherImputationStatistics"></a>}}
\if{latex}{\out{\hypertarget{method-gatherImputationStatistics}{}}}
\subsection{Method \code{gatherImputationStatistics()}}{
Gathers statistical information about missing values
that are provided by the classes public \code{gather} functions.
The information is stored within the classes instance variable \code{imputationParameter}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.missings$gatherImputationStatistics(data_df = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_df}}{The data frame to be analyzed.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$gatherInformationStatistics(data_df)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-detectImputationSites"></a>}}
\if{latex}{\out{\hypertarget{method-detectImputationSites}{}}}
\subsection{Method \code{detectImputationSites()}}{
Detects missing values within the data frame and
writes the to the instance variable \code{imputationsites}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.missings$detectImputationSites(data_df = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_df}}{The data frame to be analyzed.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{y <- x$detectImputationsites(data_df)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-imputationSiteDistribution"></a>}}
\if{latex}{\out{\hypertarget{method-imputationSiteDistribution}{}}}
\subsection{Method \code{imputationSiteDistribution()}}{
Numeric representation of the distribution of missing values within the data frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.missings$imputationSiteDistribution(data_df = "tbl_df")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_df}}{The data frame to be analyzed.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame
(tibble::tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{data \%>\%
 x$imputationSiteDistribution() \%>\%
 print()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mergeImputationSiteData"></a>}}
\if{latex}{\out{\hypertarget{method-mergeImputationSiteData}{}}}
\subsection{Method \code{mergeImputationSiteData()}}{
Merges the numeric attributes of the pguIMP data with its metadata.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.missings$mergeImputationSiteData(
  data_df = "tbl_df",
  metadata_df = "tbl_df"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_df}}{The data frame to be analyzed.
(tibble::tibble)}

\item{\code{metadata_df}}{The corresponding metadata.
(tibble::tibble)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame
(tibble::tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{data \%>\%
 x$mergeImputationSiteData() \%>\%
 print()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-imputationSiteHeatMap"></a>}}
\if{latex}{\out{\hypertarget{method-imputationSiteHeatMap}{}}}
\subsection{Method \code{imputationSiteHeatMap()}}{
Displays the distribution of missing values in form of a heatmap.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.missings$imputationSiteHeatMap()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A heatmap plot.
(ggplot2::ggplot)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$imputationSiteHeatMap() \%>\%
 show()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.missings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
