% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pguValidator.R
\name{pgu.validator}
\alias{pgu.validator}
\title{pgu.validator}
\format{
\link[R6:R6Class]{R6::R6Class} object.
}
\description{
Validates that the distribution is not significantly altered by the imputation process.
}
\details{
Takes two distributions (before and after imputation).
Performs a Wilcoxon-Mann-Whitney U test.
Performs a Kolmogorow-Smirnow test.
}
\section{Construction}{

x <- pguIMP::pgu.validator$new()
}

\examples{

## ------------------------------------------------
## Method `pgu.validator$new`
## ------------------------------------------------

x <- pguIMP:pgu.validator$new(seed = 42)

## ------------------------------------------------
## Method `pgu.validator$print`
## ------------------------------------------------

x$print()
print(x)

## ------------------------------------------------
## Method `pgu.validator$resetValidator`
## ------------------------------------------------

x$resetValidator()

## ------------------------------------------------
## Method `pgu.validator$kolmogorowTestFeature`
## ------------------------------------------------

x$kolmogorowTestFeature(org, imp, feature)

## ------------------------------------------------
## Method `pgu.validator$wilcoxonTestFeature`
## ------------------------------------------------

x$wilcoxonTestFeature(org, imp, feature)

## ------------------------------------------------
## Method `pgu.validator$centralMomentsFeature`
## ------------------------------------------------

x$centralMomentsFeature(values, feature)

## ------------------------------------------------
## Method `pgu.validator$validate`
## ------------------------------------------------

x$validate(org, imp, progress)
}
\author{
Sebastian Malkusch, \email{malkusch@med.uni-frankfurt.de}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{testStatistics_df}}{Returns the instance variable \code{testStatistics_df}.
(tibble::tibble)}

\item{\code{centralMoments_org}}{Returns the instance variable \code{centralMoments_org}
(tibble::tibble)}

\item{\code{centralMoments_imp}}{Returns the instance variable \code{centralMoments_imp}
(tibble::tibble)}

\item{\code{centralMoments_delta}}{Returns the instance variable \code{centralMoments_delta}
(tibble::tibble)}

\item{\code{features}}{Returns the instance variable \code{features}
(character)}

\item{\code{seed}}{Returns the instance variable seed
(integer)}

\item{\code{setSeed}}{Sets the instance variable seed.
(numeric)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{pgu.validator$new()}}
\item \href{#method-finalize}{\code{pgu.validator$finalize()}}
\item \href{#method-print}{\code{pgu.validator$print()}}
\item \href{#method-resetValidator}{\code{pgu.validator$resetValidator()}}
\item \href{#method-kolmogorowTestFeature}{\code{pgu.validator$kolmogorowTestFeature()}}
\item \href{#method-wilcoxonTestFeature}{\code{pgu.validator$wilcoxonTestFeature()}}
\item \href{#method-centralMomentsFeature}{\code{pgu.validator$centralMomentsFeature()}}
\item \href{#method-validate}{\code{pgu.validator$validate()}}
\item \href{#method-clone}{\code{pgu.validator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates and returns a new \code{pgu.validator} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.validator$new(seed = 42)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seed}}{Set the instance variable \code{seed}.
(integer)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{pgu.validator} object.
(pguIMP::pgu.validator)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x <- pguIMP:pgu.validator$new(seed = 42)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Clears the heap and
indicates that instance of \code{pgu.validator} is removed from heap.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.validator$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints instance variables of a \code{pgu.validator} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.validator$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
string
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$print()
print(x)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-resetValidator"></a>}}
\if{latex}{\out{\hypertarget{method-resetValidator}{}}}
\subsection{Method \code{resetValidator()}}{
Resets instance variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.validator$resetValidator()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$resetValidator()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kolmogorowTestFeature"></a>}}
\if{latex}{\out{\hypertarget{method-kolmogorowTestFeature}{}}}
\subsection{Method \code{kolmogorowTestFeature()}}{
Performs a comparison between the original and the imputated distribution of a given feature
using a two-sided Kolmorogow-Smirnow test with simulated p-vaue distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.validator$kolmogorowTestFeature(
  org = "numeric",
  imp = "numeric",
  feature = "character"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org}}{Original data to be analzed.
(numeric)}

\item{\code{imp}}{Imputed data to be analyzed.
(numeric)}

\item{\code{feature}}{Feature name of the analyzed distributions.
(character)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
One row dataframe comprising the test results.
(tibble::tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$kolmogorowTestFeature(org, imp, feature)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wilcoxonTestFeature"></a>}}
\if{latex}{\out{\hypertarget{method-wilcoxonTestFeature}{}}}
\subsection{Method \code{wilcoxonTestFeature()}}{
Performs a comparison between the original and the imputated distribution of a given feature
using a two-sided Wilcoxon/Mann-Whitney test.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.validator$wilcoxonTestFeature(
  org = "numeric",
  imp = "numeric",
  feature = "character"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org}}{Original data to be analzed.
(numeric)}

\item{\code{imp}}{Imputed data to be analyzed.
(numeric)}

\item{\code{feature}}{Feature name of the analyzed distributions.
(character)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
One row dataframe comprising the test results.
(tibble::tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$wilcoxonTestFeature(org, imp, feature)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-centralMomentsFeature"></a>}}
\if{latex}{\out{\hypertarget{method-centralMomentsFeature}{}}}
\subsection{Method \code{centralMomentsFeature()}}{
Estimates estimates the central moments of the given distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.validator$centralMomentsFeature(values = "numeric", feature = "character")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{values}}{Data to be analzed.
(numeric)}

\item{\code{feature}}{Feature name of the analyzed distributions.
(character)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
One row dataframe comprising the statistics.
(tibble::tibble)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$centralMomentsFeature(values, feature)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
Validates the feature distributions of the original and the imputated dataframe``
using a two-sided Kolmorogow-Smirnow test and a two-sided Wilcoxon/Mann-Whitney test.
The result is stored in the instance variables \code{testStatistics_df}and `distributionStatistics_df`.
Displays the progress if shiny is loaded.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.validator$validate(org = "tbl_df", imp = "tbl_df", progress = "Progress")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org}}{Original dataframe to be analzed.
(tibble::tibble)}

\item{\code{imp}}{Imputed dataframe to be analyzed.
(tibble::tibble)}

\item{\code{progress}}{If shiny is loaded, the analysis' progress is stored in this instance of the shiny Progress class.
(shiny::Progress)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{x$validate(org, imp, progress)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pgu.validator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
