\name{pguIMP-package}
\alias{pguIMP-package}
\docType{package}
\title{
\packageTitle{pguIMP}
}
\description{
\packageDescription{pguIMP}
}
\author{
\packageAuthor{pguIMP}

Maintainer: \packageMaintainer{pguIMP}
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = pguIMP)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{pguIMP}}
%\section{Documentation Index}{\packageIndices{pguIMP}}
\examples{
# simple examples of the most important functions
}
\references{
Malkusch S, Hahnefeld L, Gurke R, Lotsch J (2021).
"Visually guided preprocessing of bioanalytical laboratory data using an interactive R notebook (pguIMP)."
\emph{CPT: Pharmacometrics and Systems Pharmacology}, \strong{10}(11), 1371--1381.
\doi{10.1002/psp4.12704}
}

