% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pguIMP.R
\name{pguIMP}
\alias{pguIMP}
\title{pguIMP}
\usage{
pguIMP()
}
\value{
shiny application object
}
\description{
Reproducible cleaning of biomedical laboratory data using methods of visualization, error correction and transformation implemented as interactive R-notebooks.
}
\details{
A graphical data preprocessing toolbox, named “pguIMP”, that includes a fixed sequence of preprocessing steps to enable error-free data preprocessing interactively.
By implementing contemporary data processing methods including machine learning-based imputation procedures, the creation of corrected and cleaned bioanalytical datasets is ensured, which preserve data structures such as clusters better than resulting with classical methods.
}
\author{
Sebastian Malkusch, \email{malkusch@med.uni-frankfurt.de}
}
