% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sLogLikelihood.R
\name{sLogLikelihood}
\alias{sLogLikelihood}
\title{dLogLikelihood}
\usage{
sLogLikelihood(mu = "numeric", sigma = "numeric")
}
\arguments{
\item{mu}{The expextation value.
(numeric)}

\item{sigma}{The standard deviation.
(numeric)}
}
\value{
the bbmle::snorm results.
}
\description{
Calculates bbmle snmor function.
}
\examples{
y <- sLogLikelihood (mu=0.0, sigma=1.0)

}
\author{
Sebastian Malkusch, \email{malkusch@med.uni-frankfurt.de}
}
