% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dLogLikelihood.R
\name{dLogLikelihood}
\alias{dLogLikelihood}
\title{dLogLikelihood}
\usage{
dLogLikelihood(x = "numeric", pars = c(mu = 0, sigma = 1))
}
\arguments{
\item{x}{The x-value.
(numeric)}

\item{pars}{Numeric vector with two entries
1: The expextation value.
2: The standard deviation.
(numeric)}
}
\value{
The logLikelihood.
(numeric)
}
\description{
Calculates the logLikelihood.
}
\examples{
y <- dLogLikelihood(x=5, pars = c(mu=0.0, sigma=1.0))

}
\author{
Sebastian Malkusch, \email{malkusch@med.uni-frankfurt.de}
}
