% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pguImp.R
\name{IMPgui}
\alias{IMPgui}
\title{main}
\usage{
IMPgui()
}
\value{
shiny application object
}
\description{
Run shiny based gui of pguIMP.
}
\examples{
pguIMP::IMPgui()

}
\author{
Sebastian Malkusch, \email{malkusch@med.uni-frankfurt.de}
}
