% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importDataSet.R
\name{importDataSet}
\alias{importDataSet}
\title{importDataSet}
\usage{
importDataSet(obj = "pgu.file")
}
\arguments{
\item{obj}{Instance of the R6 class pguIMP::pgu.file.}
}
\value{
A data frame that contains the imported data
(tibble::tibble)
}
\description{
Imports a data set to the shiny 'pguIMP' web interface.
Extracts import options from a 'pguIMP::file' instance
and imports the desired record based on the passed information.
}
\author{
Sebastian Malkusch, \email{malkusch@med.uni-frankfurt.de}
}
