% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalDistribution.R
\name{normalDistribution}
\alias{normalDistribution}
\title{normalDistribution}
\usage{
normalDistribution(x = "numeric", mu = "numeric", sigma = "numeric")
}
\arguments{
\item{x}{The x-value (numeric)}

\item{mu}{The expextation value (numeric)}

\item{sigma}{The standard deviation (numeric)}
}
\value{
The probability of observing event x given mu and sigma. (numeric)
}
\description{
Probability density distribution of a normally distributed variable.
}
\details{
Calculates p(x | mu, sigma). Where p is the probability of observing an event x
given the expected value mu and the standard deviation sigma.
}
\examples{
y <- pguIMP::normalDistribution(x=5, mu=0.0, sigma=1.0)
}
\author{
Sebastian Malkusch, \email{malkusch@med.uni-frankfurt.de}
}
