% Generated by roxygen2 (disabled): do not edit by hand
\name{pgu.limitsOfQuantification}
\alias{pgu.limitsOfQuantification}
\title{pgu.limitsOfQuantification}
\format{
\link[R6:R6Class]{R6::R6Class} object.
}
\description{
Handles values in the pguIMP dataset that exceed the limits of quantification.
This object is used by the shiny based gui and is not for use in individual R-scripts!
}
\details{
more information
}
\author{
Sebastian Malkusch
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{loq}}{Returns the instance variable loq (tibble::tibble)}
\item{\code{setLoq}}{Sets the instance variable loq (tibble::tibble)}
\item{\code{outliers}}{Returns instance variable outliers (tibble::tibble)}
\item{\code{lloqSubstituteAlphabet}}{Returns the instance variable lloqSubstititeAlphabet (character)}
\item{\code{lloqSubstituteAgent}}{Returns the instance variable lloqSubstituteAgent (character)}
\item{\code{setLloqSubstituteAgent}}{Sets the instance variable lloqSubstituteAgent (character)}
\item{\code{uloqSubstituteAlphabet}}{Returns the instance variable uloqSubstititeAlphabet (character)}
\item{\code{uloqSubstituteAgent}}{Returns the instance variable uloqSubstituteAgent (character)}
\item{\code{setUloqSubstituteAgent}}{Sets the instance variable uloqSubstituteAgent (character)}
\item{\code{naHandlingAlphabet}}{Returns the instance variable naHandlingAlphabet (character)}
\item{\code{naHandlingAgent}}{Returns the instance variable naHandlingAgent (character)}
\item{\code{setNaHandlingAgent}}{Sets the instance variable naHandlingAgent (character)}
\item{\code{loqStatistics}}{Returns the instance variable loqStatistics}
}
\if{html}{\out{</div>}}
}

\section{Methods}{
The following public methods are available:

- `new()`: Creates a new pgu.limitsOfQuantification object
- `print()`: Prints instance variables  
- `reset()`: Resets the object
- `fit()`: Analyzes data for LOQ violations
- `predict()`: Applies LOQ corrections
- `attribute_lloq()`, `attribute_uloq()`: Get LOQ values
- `set_attribute_lloq()`, `set_attribute_uloq()`: Set LOQ values
- `attribute_outliers()`: Get outliers for an attribute
- `plot_loq_distribution()`: Plot LOQ statistics
- `attribute_bar_plot()`, `attribute_box_plot_with_subset()`, `attribute_plot()`: Create diagnostic plots
- `clone()`: Clone the object
}
\seealso{
Other pguIMP classes: \code{\link{pgu.missings}}
}
\keyword{datasets}
