% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transposeTibble.R
\name{transposeTibble}
\alias{transposeTibble}
\title{transposeTibble}
\usage{
transposeTibble(obj = "tbl_df")
}
\arguments{
\item{obj}{The data frame to be transposed.
(numeric)}
}
\value{
The transposed data frame.
(tibble:tibble)
}
\description{
Transposes a tibble
This object is used by the shiny based gui and is not for use in individual R-scripts!
}
\author{
Sebastian Malkusch, \email{malkusch@med.uni-frankfurt.de}
}
